requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SocketIO==5.4.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
pillow==11.0.0
pyperclip==1.9.0
python-engineio==4.10.1
python-socketio==5.11.4
qrcode==8.0
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.4
wsproto==1.2.0
